[project-test] Move back to integration test, removing docker kill invocation
