+++ src/it/scala/temple/EndpointIntegrationTest.scala
@@ -1,10 +1,10 @@
 package temple
 
 import org.scalatest.Matchers
-import temple.spec.EndpointTesterSpec
+import temple.containers.EndpointTesterSpec
 import temple.utils.FileUtils
 
-class EndpointE2ETest extends EndpointTesterSpec with Matchers {
+class EndpointIntegrationTest extends EndpointTesterSpec with Matchers {
   behavior of "temple generate"
 
   it should "correctly generate simple.temple" in {
