+++ src/it/scala/temple/containers/EndpointTesterSpec.scala
@@ -1,4 +1,4 @@
-package temple.spec
+package temple.containers
 
 import java.nio.file.Files
