+++ src/it/scala/temple/containers/SwaggerSpec.scala
@@ -16,6 +16,10 @@ abstract class SwaggerSpec extends DockerShell2HttpService(8082) with DockerTest
   // Validate a given OpenAPI specification, returning the output of swagger-cli
   def validate(openapi: String): String = {
     val json = Map("contents" -> openapi).asJson.toString()
-    Http(swaggerVerifyUrl).params(Map("openapi" -> json)).asString.body
+    Http(swaggerVerifyUrl)
+      .params(Map("openapi" -> json))
+      .timeout(connTimeoutMs = 1000, readTimeoutMs = 30000)
+      .asString
+      .body
   }
 }
